Update get_pvalues.R
Added requirement for pbkrtest if method="KR". Added function to add significance "stars" to the model output, using standard style. 
NOTE: I was not able to get the "KR" method to work on my machine; it hung up for ten minutes and I had to crash it. So I commented out the star function for "KR" and "all" methods.